requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
config==0.5.1
cssselect==1.2.0
DataRecorder==3.6.2
distro==1.9.0
DownloadKit==2.0.6
DrissionPage==4.1.0.12
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fake-useragent==1.5.1
filelock==3.16.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
jiter==0.8.0
lxml==5.3.0
lxml_html_clean==0.4.1
markdown2==2.5.1
openai==1.55.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
parse==1.20.2
playwright==1.46.0
pluggy==1.5.0
psutil==6.1.0
pydantic==2.10.2
pydantic_core==2.27.1
pyee==11.1.0
PyMuPDF==1.24.14
pymupdf4llm==0.0.17
PyPDF2==3.0.1
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0
requests-html==0.10.0
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tldextract==5.1.3
tomli==2.2.1
tqdm==4.67.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.20
w3lib==2.2.1
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.21.0
